# Optimize core*.test_avx to run faster.