… test/sse/test_avx.cpp