Give the same no-inline treatment to AVX test than the earlier SSE 4.1 test.

Before:

```
test_avx (test_core.core0.test_avx) ... ok (14.30s)
test_avx (test_core.core1.test_avx) ... ok (15.00s)
test_avx (test_core.core_2gb.test_avx) ... ok (14.43s)
test_avx (test_core.core2.test_avx) ... ok (17.92s)
test_avx (test_core.corez.test_avx) ... ok (42.94s)
test_avx (test_core.cores.test_avx) ... ok (65.45s)
test_avx (test_core.core3.test_avx) ... ok (1026.48s)
```

After:

```
test_avx (test_core.core0.test_avx) ... ok (14.70s)
test_avx (test_core.core1.test_avx) ... ok (13.78s)
test_avx (test_core.core_2gb.test_avx) ... ok (15.00s)
test_avx (test_core.core2.test_avx) ... ok (15.95s)
test_avx (test_core.corez.test_avx) ... ok (35.31s)
test_avx (test_core.cores.test_avx) ... ok (39.06s)
test_avx (test_core.core3.test_avx) ... ok (45.72s)
```